# backend/docs/swagger.yaml
basePath: /
definitions:
  models.CDNLine:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      id:
        type: integer
      name:
        type: string
      provider:
        $ref: '#/definitions/models.CDNProvider'
      provider_id:
        type: integer
      updated_at:
        type: string
    type: object
  models.CDNProvider:
    properties:
      code:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  models.CreateCDNLineRequest:
    properties:
      display_name:
        maxLength: 255
        minLength: 1
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      provider_id:
        type: integer
    required:
    - display_name
    - name
    - provider_id
    type: object
  models.CreateCDNProviderRequest:
    properties:
      code:
        maxLength: 100
        minLength: 1
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - code
    - name
    type: object
...
  models.UpdateCDNLineRequest:
    properties:
      display_name:
        maxLength: 255
        minLength: 1
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      provider_id:
        type: integer
    required:
    - display_name
    - name
    - provider_id
    type: object
  models.UpdateCDNProviderRequest:
    properties:
      code:
        maxLength: 100
        minLength: 1
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - code
    - name
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Video Manager API Documentation
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Video Manager API
  version: "1.0"
paths:
...
  /api/cdn-lines:
    get:
      description: Retrieve all CDN lines, optionally filtered by provider ID
      parameters:
      - description: Filter by provider ID
        in: query
        name: provider_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.CDNLine'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get all CDN lines
      tags:
      - cdn-lines
    post:
      consumes:
      - application/json
      description: Create a new CDN line with the provided information
      parameters:
      - description: CDN Line information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateCDNLineRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.CDNLine'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create a new CDN line
      tags:
      - cdn-lines
  /api/cdn-lines/{id}:
    delete:
      description: Delete a CDN line by its ID. This will also delete associated video
        stream endpoints.
      parameters:
      - description: Line ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete a CDN line
      tags:
      - cdn-lines
    get:
      description: Retrieve a CDN line by its ID
      parameters:
      - description: Line ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.CDNLine'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get CDN line by ID
      tags:
      - cdn-lines
    put:
      consumes:
      - application/json
      description: Update an existing CDN line by its ID
      parameters:
      - description: Line ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated CDN Line information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCDNLineRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.CDNLine'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update a CDN line
      tags:
      - cdn-lines
  /api/cdn-providers:
    get:
      description: Retrieve all CDN providers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.CDNProvider'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get all CDN providers
      tags:
      - cdn-providers
    post:
      consumes:
      - application/json
      description: Create a new CDN provider with the provided information
      parameters:
      - description: CDN Provider information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateCDNProviderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.CDNProvider'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create a new CDN provider
      tags:
      - cdn-providers
  /api/cdn-providers/{id}:
    delete:
      description: Delete a CDN provider by its ID. This will also delete associated
        CDN lines and video stream endpoints.
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete a CDN provider
      tags:
      - cdn-providers
    get:
      description: Retrieve a CDN provider by its ID
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.CDNProvider'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get CDN provider by ID
      tags:
      - cdn-providers
    put:
      consumes:
      - application/json
      description: Update an existing CDN provider by its ID
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated CDN Provider information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCDNProviderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.CDNProvider'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update a CDN provider
      tags:
      - cdn-providers
  /api/cdn-providers/{id}/lines:
    get:
      description: Retrieve all CDN lines associated with a specific provider
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.CDNLine'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get CDN lines by provider
      tags:
      - cdn-providers
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"